Major Refactoring & Route Selection Set-Up
I refactored my files so that they are part of a logically structured filesystem with folders.

I also set up the route selection framework. The button to create a route in the user interface now initiates a call to the routeCalculator() function with the necessary data.

I have also written pseudocode for route selection on paper, which I will begin implementing soon.